FP-10: Moved widgets to correct folder
