+++ assets/pages/Recipes/AddRecipe.tsx
@@ -1,7 +1,7 @@
-import { InputWidget } from "@/components/form2/InputWidget"
-import { LabelledFormWidget } from "@/components/form2/LabelledFormWidget"
-import { SliderWidget } from "@/components/form2/SliderWidget"
-import { TextareaWidget } from "@/components/form2/TextareaWidget"
+import { InputWidget } from "@/components/form/InputWidget"
+import { LabelledFormWidget } from "@/components/form/LabelledFormWidget"
+import { SliderWidget } from "@/components/form/SliderWidget"
+import { TextareaWidget } from "@/components/form/TextareaWidget"
 import Button from '@/components/ui/Buttons/Button'
 import Card from '@/components/ui/Card'
 import Spacer from '@/components/ui/Spacer'
